requirements.txt
playwright>=1.45.1
google-generativeai>=0.7.2
pyyaml>=6.0.2